+++ plugins/woocommerce-admin/packages/components/src/list/index.js
@@ -115,3 +115,4 @@ export default List;
 
 export { ExperimentalListItem } from './experimental-list-item';
 export { ExperimentalList } from './experimental-list';
+export { ExperimentalCollapsibleList } from './collapsible-list';