+++ plugins/woocommerce-admin/packages/components/src/list/test/index.js
@@ -7,13 +7,21 @@ jest.mock( '../list-item', () => ( {
 /**
  * External dependencies
  */
-import { render, screen } from '@testing-library/react';
+import {
+	render,
+	screen,
+	waitForElementToBeRemoved,
+} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
 /**
  * Internal dependencies
  */
-import List, { ExperimentalList, ExperimentalListItem } from '../index';
+import List, {
+	ExperimentalList,
+	ExperimentalListItem,
+	ExperimentalCollapsibleList,
+} from '../index';
 import { handleKeyDown } from '../list-item';
 
 describe( 'List', () => {
@@ -143,6 +151,124 @@ describe( 'List', () => {
 				expect( item ).toBeInTheDocument();
 				expect( item ).toHaveAttribute( 'role', 'button' );
 				expect( item ).toHaveAttribute( 'tabindex', '0' );
+			} );
+		} );
+
+		describe( 'ExperimentalListItemCollapse', () => {
+			it( 'should not render its children intially, but an extra list footer with show text', () => {
+				const { container } = render(
+					<ExperimentalCollapsibleList
+						collapseLabel="Show less"
+						expandLabel="Show more items"
+					>
+						<div>Test</div>
+					</ExperimentalCollapsibleList>
+				);
+
+				expect( container ).not.toHaveTextContent( 'Test' );
+				expect( container ).toHaveTextContent( 'Show more items' );
+			} );
+
+			it( 'should render list items when footer is clicked and trigger onExpand', () => {
+				const onExpand = jest.fn();
+				const onCollapse = jest.fn();
+				const { container } = render(
+					<ExperimentalCollapsibleList
+						collapseLabel="Show less"
+						expandLabel="Show more items"
+						onExpand={ onExpand }
+						onCollapse={ onCollapse }
+					>
+						<div>Test</div>
+						<div>Test 2</div>
+					</ExperimentalCollapsibleList>
+				);
+
+				const listItem = container.querySelector(
+					'.list-item-collapse'
+				);
+
+				userEvent.click( listItem );
+				expect( container ).toHaveTextContent( 'Test' );
+				expect( container ).toHaveTextContent( 'Test 2' );
+				expect( container ).not.toHaveTextContent( 'Show more items' );
+				expect( container ).toHaveTextContent( 'Show less' );
+				expect( onExpand ).toHaveBeenCalled();
+				expect( onCollapse ).not.toHaveBeenCalled();
+			} );
+
+			it( 'should render minimum children if minChildrenToShow is set and show the rest on expand', () => {
+				const onExpand = jest.fn();
+				const onCollapse = jest.fn();
+				const { container } = render(
+					<ExperimentalCollapsibleList
+						collapseLabel="Show less"
+						expandLabel="Show more items"
+						onExpand={ onExpand }
+						onCollapse={ onCollapse }
+						show={ 2 }
+					>
+						<div>Test</div>
+						<div>Test 2</div>
+						<div>Test 3</div>
+						<div>Test 4</div>
+					</ExperimentalCollapsibleList>
+				);
+
+				expect( container ).toHaveTextContent( 'Test' );
+				expect( container ).toHaveTextContent( 'Test 2' );
+				expect( container ).not.toHaveTextContent( 'Test 3' );
+				expect( container ).not.toHaveTextContent( 'Test 4' );
+				const listItem = container.querySelector(
+					'.list-item-collapse'
+				);
+
+				userEvent.click( listItem );
+				expect( container ).toHaveTextContent( 'Test' );
+				expect( container ).toHaveTextContent( 'Test 2' );
+				expect( container ).toHaveTextContent( 'Test 3' );
+				expect( container ).toHaveTextContent( 'Test 4' );
+				expect( container ).not.toHaveTextContent( 'Show more items' );
+				expect( container ).toHaveTextContent( 'Show less' );
+				expect( onExpand ).toHaveBeenCalled();
+				expect( onCollapse ).not.toHaveBeenCalled();
+			} );
+
+			it( 'should correctly toggle the list', async () => {
+				const onExpand = jest.fn();
+				const onCollapse = jest.fn();
+				const { container } = render(
+					<ExperimentalCollapsibleList
+						collapseLabel="Show less"
+						expandLabel="Show more items"
+						onExpand={ onExpand }
+						onCollapse={ onCollapse }
+					>
+						<div id="test">Test</div>
+						<div>Test 2</div>
+					</ExperimentalCollapsibleList>
+				);
+
+				let listItem = container.querySelector( '.list-item-collapse' );
+
+				userEvent.click( listItem );
+				expect( container ).toHaveTextContent( 'Test' );
+				expect( container ).toHaveTextContent( 'Test 2' );
+				expect( container ).not.toHaveTextContent( 'Show more items' );
+				expect( container ).toHaveTextContent( 'Show less' );
+
+				listItem = container.querySelector( '.list-item-collapse' );
+
+				userEvent.click( listItem );
+				await waitForElementToBeRemoved(
+					container.querySelector( '#test' )
+				);
+				expect( container ).not.toHaveTextContent( 'Test' );
+				expect( container ).not.toHaveTextContent( 'Test 2' );
+				expect( container ).toHaveTextContent( 'Show more items' );
+				expect( container ).not.toHaveTextContent( 'Show less' );
+				expect( onExpand ).toHaveBeenCalledTimes( 1 );
+				expect( onCollapse ).toHaveBeenCalledTimes( 1 );
 			} );
 		} );
 	} );