+++ plugins/woocommerce-admin/packages/components/CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+-   Add new (experimental) collapsible list item to collapse list items. #6869
 -   SelectControl: fix display of multiple selections without inline tags. #6862
 -   Add new (experimental) list, and add depreciation notice for the current list. #6787
 -   Force `<SearchListItem>` form elements id to be unique. #6871